feat[PopUp]: menambahkan pop up pada blur
