+++ src/components/academy/CourseSection.js
@@ -1,8 +1,7 @@
 // Import Packages
 "use client";
 
-import React from "react";
-import { FaBriefcase } from "react-icons/fa";
+import React, { useState } from "react";
 import Image from "next/image";
 import { Swiper, SwiperSlide } from "swiper/react";
 import Link from "next/link";
@@ -10,12 +9,11 @@ import "swiper/css";
 import "swiper/css/effect-coverflow";
 import "swiper/css/pagination";
 import "swiper/css/navigation";
-
-import { EffectCoverflow, Pagination, Navigation } from "swiper";
+import "swiper/css";
+import { Pagination } from "swiper/modules";
 
 // Import Configs
 import { createBackground } from "@/config/Functions";
-import Slider from "./Slider";
 import Blur from "../global/Blur";
 
 export default function CourseSection() {
@@ -32,6 +30,20 @@ export default function CourseSection() {
       href: "academy/",
     },
   ];
+  // Data
+  const data = {
+    image: [
+      "/img/academy/dummyCardImage.png",
+      "/img/academy/dummyCardImage.png",
+      "/img/academy/dummyCardImage.png",
+      "/img/academy/dummyCardImage.png",
+    ],
+    module: ["Module 1", "Module 2", "Module 3", "Module 4"],
+    title: ["Consulting 101", "Business Analysis Framework", "TBA", "TBA"],
+    link: ["academy/module1", "academy/", "academy/", "academy/"],
+  };
+
+  const [swiper, setSwiper] = useState();
   return (
     <section className="relative">
       {/* Background */}
@@ -54,7 +66,74 @@ export default function CourseSection() {
               Start Your Course !
             </h1>
             <div className="w-[90vw]">
-              <Slider />
+              <div className="relative">
+                <div onClick={() => swiper.slidePrev()}>
+                  <Image
+                    src="/img/academy/arrow_back.png"
+                    alt="Previous Arrow"
+                    className="absolute bottom-[0vh] left-[41vw] z-[20] h-[2.6vw] w-[1.45vw]"
+                    width={2000}
+                    height={2000}
+                    draggable="false"
+                  />
+                </div>
+                <div onClick={() => swiper.slideNext()}>
+                  <Image
+                    src="/img/academy/arrow_forward.png"
+                    alt="Previous Arrow"
+                    className="absolute bottom-[0vh] right-[41vw] z-[20] h-[2.6vw] w-[1.45vw]"
+                    width={2000}
+                    height={2000}
+                    draggable="false"
+                  />
+                </div>
+                <Swiper
+                  height={720}
+                  spaceBetween={10}
+                  slidesPerView={2}
+                  loop={false}
+                  className="z-10"
+                  onSwiper={setSwiper}
+                  modules={[Pagination]}
+                  pagination={{ clickable: true }}
+                >
+                  {data.image.map(function (item, index) {
+                    return (
+                      <SwiperSlide key={item}>
+                        <div className="flex gap-x-[2vw]">
+                          <Blur
+                            isBlur={index !== 0}
+                            className="mb-[5vw] h-[25.313vw] w-[42.969vw] rounded-[1.6vw] bg-white shadow-xl"
+                          >
+                            <Image
+                              src={data.image[index]}
+                              alt="background"
+                              width={2000}
+                              height={2000}
+                              className="inset-0 mx-[1vw] my-[1vw] h-[14.01vw] w-[40.99vw] object-cover max-lg:hidden"
+                            />
+                            <div className="mt-[1.5vw] text-[1vw]/[1.7vw]">
+                              <h1 className="ml-[1vw] font-avenirHeavy text-[1.302vw]">
+                                {data.module[index]}:
+                              </h1>
+                              <h1 className="ml-[1vw] font-avenirHeavy text-[1.823vw] text-[#58B9D1]">
+                                {data.title[index]}
+                              </h1>
+                            </div>
+
+                            <Link
+                              href={data.link[index]}
+                              className="mx-[1.3vw] mt-[1.7vw] flex h-[2.917vw] w-[39.948vw] items-center justify-center rounded-[1.3vw] border border-black bg-white text-center text-[1.563vw] transition-all duration-700 ease-in-out hover:scale-[102%] hover:bg-[#5AB0BB]/20"
+                            >
+                              See Details
+                            </Link>
+                          </Blur>
+                        </div>
+                      </SwiperSlide>
+                    );
+                  })}
+                </Swiper>
+              </div>
             </div>
           </div>
         </div>
@@ -76,6 +155,7 @@ export default function CourseSection() {
               <>
                 <Blur
                   key={`${item + index}`}
+                  isBlur={index !== 0}
                   className="z-[3] my-[2vw] h-[51.389vw] w-[87.222vw] rounded-[1.6vw] bg-white shadow-xl"
                 >
                   <Image
